+++ lib/recipes/python/recipe.rb
@@ -30,7 +30,7 @@ PYTHON_RECIPE = Kitchen::BookRecipe.new(book_short_name: :python) do |doc, _reso
       BakeInjectedExerciseQuestion.v1(
         question: question,
         number: question.count_in(:page),
-        options: { only_number_solution: false, detailed_solution: true }
+        options: { only_number_solution: false }
       )
     end
 
@@ -76,7 +76,7 @@ PYTHON_RECIPE = Kitchen::BookRecipe.new(book_short_name: :python) do |doc, _reso
       BakeInjectedExerciseQuestion.v1(
         question: question,
         number: question.count_in(:page),
-        options: { only_number_solution: false, add_dot: true, detailed_solution: true }
+        options: { only_number_solution: false, add_dot: true }
       )
     end