+++ spec/kitchen_spec/directions/bake_injected_exercise/bake_injected_exercise_spec.rb
@@ -108,12 +108,26 @@ RSpec.describe Kitchen::Directions::BakeInjectedExercise do
 
   context 'when alphabetical_multiparts is true' do
     it 'bakes an alphabetical multipart question' do
-      described_class.v1(exercise: injected_exercise_full_with_multipart_questions_with_solutions, alphabetical_multiparts: true)
+      described_class.v1(
+        exercise: injected_exercise_full_with_multipart_questions_with_solutions,
+        options: { alphabetical_multiparts: true, list_type: 'lower-alpha', add_brackets: true }
+      )
+      expect(injected_exercise_full_with_multipart_questions_with_solutions).to match_snapshot_auto
+    end
+
+    it 'bakes an alphabetical multipart question with different numbering' do
+      described_class.v1(
+        exercise: injected_exercise_full_with_multipart_questions_with_solutions,
+        options: { alphabetical_multiparts: true, list_type: 'upper-alpha' }
+      )
       expect(injected_exercise_full_with_multipart_questions_with_solutions).to match_snapshot_auto
     end
 
     it 'leaves an exercise without a stimulus alone' do
-      described_class.v1(exercise: injected_exercise_none, alphabetical_multiparts: true)
+      described_class.v1(
+        exercise: injected_exercise_none,
+        options: { alphabetical_multiparts: true, list_type: 'lower-alpha' }
+      )
       expect(injected_exercise_none).to match_snapshot_auto
     end
   end