+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Refactor custom `Matcher`s to be more readable
 * Create test & test data for web pipeline
+* Add attribute `lang` to index foreign terms
+
+## [v2.2.0] - 2023-09-22
+
+* Fix `V4` of `BakeReferences` and apply it to `nursing-external`
+* Add options for multipart questions in `BakeInjectedExercise`
+* Revert changes made for `detailed solution` in `python`
 * Refactor: Move baking logic from recipe files to centralized executable
   * Change `bake` files to `recipe.rb`
 * Refactor: Take out `Strategy` architecture, move logic inline